videoApi,save 연결
