+++ src/pages/login/index.tsx
@@ -12,6 +12,7 @@ const Login = () => {
     )
       router.push(`https://kauth.kakao.com/oauth/authorize?client_id=${process.env.NEXT_PUBLIC_KAKAO_CLIENT_ID}&redirect_uri=${process.env.NEXT_PUBLIC_KAKAO_REDIRECT_URI}&response_type=code
     `);
+    console.log('a');
   };
 
   return (
